hot fix result mapping

+++ src/main/java/com/higgsup/xshop/dto/ProductDTO.java
@@ -18,5 +18,5 @@ public class ProductDTO {
 
   private Float avgRating;
 
-  private Long totalRating;
+  private Integer totalRating;
 }
